+++ api/database/migrations/1534883828549_payment_schema.js
@@ -10,7 +10,7 @@ class PaymentSchema extends Schema {
       table.integer('pupil_id').unsigned().references('id').inTable('pupils').onDelete('cascade')
       table.integer('paidBy').unsigned().references('id').inTable('gurdians').onDelete('cascade')
       table.integer('bank_account_id').unsigned().references('id').inTable('bank_accounts').onDelete('cascade')
-      table.integer('amount')
+      table.decimal('amount')
       table.boolean('receipted').defaultTo(false)
       table.boolean('fee').defaultTo(true)
     })
